added logout function, moved models to the same folds
