+++ src/app/header/header.component.html
@@ -9,7 +9,7 @@
       <li class="nav-item active">
 
         <!-- added login functionality - TN -->
-        <a class = "nav-link" routerLink="/login">Login</a> 
+        <a  *ngIf="!username" class = "nav-link" routerLink="/login">Login</a> 
       </li>
       <li class="nav-item">
         <a class = "nav-link" routerLink="/forums">Forums</a>
@@ -19,6 +19,9 @@
       </li>
       <li class="nav-item">
         <a class = "nav-link" routerLink="/forums/post">Post</a>
+      </li>
+      <li>
+      <a class = "nav-link" routerLink="/" (click)="logout()">Logout</a> 
       </li>
     </ul>
   </div>
